requirements.txt
aiohttp==3.12.15
aiofiles==24.1.0
miservice==2.3.0
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
PyJWT==2.10.1
PyYAML==6.0.2
cachetools==6.1.0